+++ BackEnd/payment_service/src/main/java/iuh/fit/se/payment_service/dto/RevenueDTO.java
@@ -3,9 +3,9 @@ package iuh.fit.se.payment_service.dto;
 import java.math.BigDecimal;
 
 public class RevenueDTO {
-    private BigDecimal totalRevenue;
+    private Double totalRevenue;
     private Long totalTransactions;
 
-    public RevenueDTO(BigDecimal total, long size) {
+    public RevenueDTO(Double total, long size) {
     }
 }
